pyproject.toml
[tool.poetry]
name = "nonebot-plugin-groups-aichat"
version = "0.1.0"
description = ""
authors = ["KarisAya <1048827424@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
nonebot2 = "^2.4.1"
nonebot-plugin-clovers = {version = "^0.2.0", extras = ["onebot"]}
clovers-aichat = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
